feature(build-infra): Build project multiple times
The project needs  to be built at  most 'num_sources-1'  times for all
dependencies to be  solved because they won't be in the right order in
this setup. In this version  of the build script, the library is built
silently (without  error/debug output) 'num_sources-1' times  and then
once again with all outputs.